elki generator now sets labels

+++ clustergenerator/src/main/java/clusterproject/program/Generator/ELKIGenerator.java
@@ -4,19 +4,24 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.PrintWriter;
 import java.io.UnsupportedEncodingException;
+import java.util.SortedSet;
+import java.util.regex.Pattern;
 
 import javax.swing.JPanel;
 
 import clusterproject.data.PointContainer;
 import clusterproject.program.Generator.Panel.ELKIOptions;
+import de.lmu.ifi.dbs.elki.data.ClassLabel;
 import de.lmu.ifi.dbs.elki.data.NumberVector;
 import de.lmu.ifi.dbs.elki.data.type.TypeUtil;
 import de.lmu.ifi.dbs.elki.database.Database;
 import de.lmu.ifi.dbs.elki.database.StaticArrayDatabase;
-import de.lmu.ifi.dbs.elki.database.ids.DBIDIter;
+import de.lmu.ifi.dbs.elki.database.ids.ArrayModifiableDBIDs;
+import de.lmu.ifi.dbs.elki.database.ids.DBIDArrayMIter;
 import de.lmu.ifi.dbs.elki.database.relation.Relation;
 import de.lmu.ifi.dbs.elki.datasource.GeneratorXMLDatabaseConnection;
 import de.lmu.ifi.dbs.elki.utilities.ClassGenericsUtil;
+import de.lmu.ifi.dbs.elki.utilities.DatabaseUtil;
 import de.lmu.ifi.dbs.elki.utilities.datastructures.arraylike.ArrayLikeUtil;
 import de.lmu.ifi.dbs.elki.utilities.exceptions.AbortException;
 import de.lmu.ifi.dbs.elki.utilities.optionhandling.parameterization.ListParameterization;
@@ -46,7 +51,8 @@ public class ELKIGenerator implements IGenerator {
 
 	@Override
 	public boolean generate(PointContainer container) {
-		final String tempInName = "test.xml";
+		boolean debug = false;
+		final String tempInName = "temp.xml";
 		PrintWriter writer = null;
 		try {
 			writer = new PrintWriter(tempInName, "UTF-8");
@@ -75,37 +81,46 @@ public class ELKIGenerator implements IGenerator {
 		} catch (final AbortException e) {
 			// TODO: handle exception
 			final File file1 = new File(tempInName);
-			file1.delete();
+			if (!debug)
+				file1.delete();
 			return false;
 		}
-		// Check the relation has the expected size:
 
 		final Relation<NumberVector> rel = db.getRelation(TypeUtil.NUMBER_VECTOR_FIELD);
 
 		final int newDim = ArrayLikeUtil.toPrimitiveDoubleArray(rel.get(rel.getDBIDs().iter())).length;
 		if (newDim != container.getDim() && !optionsPanel.replacePoints()) {
 			final File file1 = new File(tempInName);
-			file1.delete();
+			if (!debug)
+				file1.delete();
 			return false;// TODO set error
 		}
 
 		if (optionsPanel.replacePoints()) {
 			container.empty();
 		}
+		SortedSet<ClassLabel> labels = DatabaseUtil.getClassLabels(db);
 
-		for (final DBIDIter it = rel.getDBIDs().iter(); it.valid(); it.advance()) {
-
-			// To get the vector use:
-			final NumberVector v = rel.get(it);
-			container.addPoint(ArrayLikeUtil.toPrimitiveDoubleArray(v));
+		int cid = 0;
+		if (optionsPanel.replacePoints())
+			container.setUpClusters();
+		for (ClassLabel lbl : labels) {
+			ArrayModifiableDBIDs amdbids = DatabaseUtil.getObjectsByLabelMatch(db, Pattern.compile(lbl.toString()));
+			for (DBIDArrayMIter dbamiter = amdbids.iter(); dbamiter.valid(); dbamiter.advance()) {
+				final NumberVector v = rel.get(dbamiter);
+				container.addPoint(ArrayLikeUtil.toPrimitiveDoubleArray(v));
+				if (optionsPanel.replacePoints())
+					container.getClusterInformation().addClusterID(cid);
+			}
+			cid++;
 		}
-
 		if (optionsPanel.replacePoints()) {
 			container.rebuild();
 		}
 
 		final File file1 = new File(tempInName);
-		file1.delete();
+		if (!debug)
+			file1.delete();
 		return true;
 	}
